Omes CLI Docker image
