+++ cmd/build_worker_image.go
@@ -4,15 +4,12 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/temporalio/omes/cmd/cmdoptions"
-	"go.uber.org/zap"
 	"golang.org/x/mod/semver"
 )
 
@@ -51,37 +48,20 @@ func buildPushWorkerImageCmd() *cobra.Command {
 }
 
 type workerImageBuilder struct {
-	logger         *zap.SugaredLogger
-	tagAsLatest    bool
-	platforms      []string
-	imageName      string
-	repoPrefix     string
-	dryRun         bool
-	saveImage      string
-	tags           []string
-	labels         []string
-	loggingOptions cmdoptions.LoggingOptions
-	sdkOptions     cmdoptions.SdkOptions
+	baseImageBuilder
+	sdkOptions cmdoptions.SdkOptions
 }
 
 func (b *workerImageBuilder) addCLIFlags(fs *pflag.FlagSet) {
 	b.sdkOptions.AddCLIFlags(fs)
-	fs.BoolVar(&b.tagAsLatest, "tag-as-latest", false,
-		"If set, tag the image as latest in addition to the omes commit sha tag")
-	fs.StringSliceVar(&b.platforms, "platform", []string{"amd64"}, "Platforms for use in docker build --platform")
-	fs.StringVar(&b.imageName, "image-name", "omes", "Name of the image to build")
-	fs.StringVar(&b.repoPrefix, "repo-prefix", "", "Repository prefix (e.g., 'temporaliotest'). If empty, no prefix is used.")
-	fs.BoolVar(&b.dryRun, "dry-run", false, "If set, just print the commands that would run but do not run them")
-	fs.StringSliceVar(&b.tags, "image-tag", nil, "Additional tags to add to the image")
-	fs.StringSliceVar(&b.labels, "image-label", nil, "Additional labels to add to the image")
-	fs.StringVar(&b.saveImage, "save-image", "", "If set, will run `docker save` on the produced image, saving it to the provided path")
-	b.loggingOptions.AddCLIFlags(fs)
+	b.addBaseCLIFlags(fs)
 }
 
 func (b *workerImageBuilder) build(ctx context.Context, allowPush bool) error {
 	b.logger = b.loggingOptions.MustCreateLogger()
 	lang := b.sdkOptions.Language.String()
 	version := b.sdkOptions.Version
+
 	// At some point we probably want to replace this with a meaningful version of omes itself
 	omesVersion, err := getCurrentCommitSha(ctx)
 	if err != nil {
@@ -101,12 +81,15 @@ func (b *workerImageBuilder) build(ctx context.Context, allowPush bool) error {
 		} else if !filepath.IsLocal(version) {
 			return fmt.Errorf("invalid path version: must be beneath this dir")
 		}
+
 		// Dockerfile copies entire version path to ./repo
 		buildArgs = append(buildArgs, "SDK_VERSION=./repo", "SDK_DIR="+version)
 	} else {
 		buildArgs = append(buildArgs, "SDK_VERSION="+version)
+
 		// Add label for version
 		b.addLabelIfNotPresent("io.temporal.sdk.version", version)
+
 		// Check for valid version
 		versionToCheck := version
 		if !strings.HasPrefix(version, "v") {
@@ -115,6 +98,7 @@ func (b *workerImageBuilder) build(ctx context.Context, allowPush bool) error {
 		if semver.Canonical(versionToCheck) == "" {
 			return fmt.Errorf("expected valid semver")
 		}
+
 		// Add tag for lang-fullsemver without leading "v". We are intentionally
 		// including semver build metadata.
 		langTagComponent := lang + "-" + strings.TrimPrefix(version, "v")
@@ -125,116 +109,29 @@ func (b *workerImageBuilder) build(ctx context.Context, allowPush bool) error {
 			b.tags = append(b.tags, lang+"-latest")
 		}
 	}
+	imageTagsForPublish := b.generateImageTags()
 
-	// Setup additional labels
-	gitRef, err := gitRef(ctx, ".git")
+	// Set OCI labels
+	err = b.addDefaultLabels(ctx, omesVersion, "Load testing for "+lang)
 	if err != nil {
 		return err
 	}
-	b.addLabelIfNotPresent("org.opencontainers.image.created", time.Now().UTC().Format(time.RFC3339))
-	b.addLabelIfNotPresent("org.opencontainers.image.source", "https://github.com/temporalio/omes")
-	b.addLabelIfNotPresent("org.opencontainers.image.vendor", "Temporal Technologies Inc.")
-	b.addLabelIfNotPresent("org.opencontainers.image.authors", "Temporal SDK team <sdk-team@temporal.io>")
-	b.addLabelIfNotPresent("org.opencontainers.image.licenses", "MIT")
-	b.addLabelIfNotPresent("org.opencontainers.image.revision", gitRef)
-	b.addLabelIfNotPresent("org.opencontainers.image.title", "Load testing for "+lang)
-	b.addLabelIfNotPresent("org.opencontainers.image.documentation", "See README at https://github.com/temporalio/omes")
 	b.addLabelIfNotPresent("io.temporal.sdk.name", lang)
 	b.addLabelIfNotPresent("io.temporal.sdk.version", version)
-	b.addLabelIfNotPresent("io.temporal.omes.githash", omesVersion)
 
-	// Prepare docker command args
-	args := []string{
-		"buildx",
-		"build",
-		"--pull",
-		"--file", "dockerfiles/" + lang + ".Dockerfile",
-		"--platform", strings.Join(b.platforms, ","),
-	}
-	if len(b.platforms) > 1 {
+	// Build docker command args
+	args, err := b.buildDockerArgs("dockerfiles/"+lang+".Dockerfile", allowPush, buildArgs)
+	if err != nil {
 		if !allowPush {
 			return fmt.Errorf("multi-platform builds require pushing to registry. Use build-push-worker-image command instead")
 		}
-		// Multi-platform builds can't be loaded into local Docker, they must be pushed
-		args = append(args, "--push")
-	} else {
-		// Single-platform builds can be loaded into local Docker
-		args = append(args, "--load")
-	}
-	var imageTagsForPublish []string
-	for _, tag := range b.tags {
-		tagVal := fmt.Sprintf("%s:%s", b.imageName, tag)
-		if b.repoPrefix != "" {
-			tagVal = fmt.Sprintf("%s/%s", b.repoPrefix, tagVal)
-		}
-		args = append(args, "--tag", tagVal)
-		imageTagsForPublish = append(imageTagsForPublish, tagVal)
-	}
-	for _, label := range b.labels {
-		args = append(args, "--label", label)
-	}
-	for _, arg := range buildArgs {
-		args = append(args, "--build-arg", arg)
-	}
-	args = append(args, repoDir())
-	b.logger.Infof("Running: docker %v", strings.Join(args, " "))
-	if b.dryRun {
-		return nil
+		return err
 	}
 
-	// Write all the produced image tags to an env var so that the GH workflow can later use it
-	// to publish them.
-	err = writeGitHubEnv("FEATURES_BUILT_IMAGE_TAGS", strings.Join(imageTagsForPublish, ";"))
+	err = b.executeDockerBuild(ctx, args, imageTagsForPublish)
 	if err != nil {
-		return fmt.Errorf("writing image tags to github env failed: %s", err)
+		return err
 	}
 
-	cmd := exec.CommandContext(ctx, "docker", args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err = cmd.Run()
-	if err != nil {
-		return fmt.Errorf("failed building image: %w", err)
-	}
-
-	if b.saveImage != "" {
-		err = writeGitHubEnv("SAVED_IMAGE_TAG", imageTagsForPublish[0])
-		if err != nil {
-			return fmt.Errorf("writing image tags to github env failed: %s", err)
-		}
-		// For multi-platform builds, buildx doesn't load images into local Docker
-		// so we can't save them. Only single-platform builds can be saved.
-		if len(b.platforms) > 1 {
-			b.logger.Warn("Image saving is not supported for multi-platform builds. Skipping save step.")
-		} else {
-			saveCmd := exec.CommandContext(ctx, "docker", "save", "-o", b.saveImage, imageTagsForPublish[0])
-			saveCmd.Stdout = os.Stdout
-			saveCmd.Stderr = os.Stderr
-			err = saveCmd.Run()
-			if err != nil {
-				return fmt.Errorf("failed saving image: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (b *workerImageBuilder) addLabelIfNotPresent(key, value string) {
-	for _, label := range b.labels {
-		if strings.HasPrefix(label, key+"=") {
-			return
-		}
-	}
-	b.labels = append(b.labels, key+"="+value)
-}
-
-func gitRef(ctx context.Context, gitDir string) (string, error) {
-	cmd := exec.Command("git", "--git-dir", gitDir, "rev-parse", "HEAD")
-	cmd.Stderr = os.Stderr
-	out, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("failed getting git ref: %w", err)
-	}
-	return strings.TrimRight(string(out), "\r\n"), nil
+	return b.handleImageSave(ctx, imageTagsForPublish)
 }
